Rename lg_thinq domain name
