+++ homeassistant/brands/lg.json
@@ -1,5 +1,5 @@
 {
   "domain": "lg",
   "name": "LG",
-  "integrations": ["lg_netcast", "lg_soundbar", "webostv"]
+  "integrations": ["lg_netcast", "lg_thinq", "lg_soundbar", "webostv"]
 }
