+++ homeassistant/components/lg_thinq/coordinator.py
