+++ homeassistant/components/lg_thinq/entity.py
