+++ homeassistant/components/lg_thinq/icons.json
