+++ homeassistant/components/lg_thinq/strings.json
