+++ homeassistant/components/lg_thinq/switch.py
