+++ requirements_test_all.txt
@@ -2198,7 +2198,7 @@ thermobeacon-ble==0.7.0
 # homeassistant.components.thermopro
 thermopro-ble==0.10.0
 
-# homeassistant.components.lgthinq
+# homeassistant.components.lg_thinq
 thinqconnect==0.9.5
 
 # homeassistant.components.tilt_ble
