+++ tests/components/lg_thinq/const.py
